Add `bashsimplecurses` submodule
`bashsimplecurses` provides the code used to create our windows and
widgets. Bring it in as a submodule so we can pull updates from upstream
as they come in. Prevents the user from needing to install an extra
dependency as it will be included as part of this main repo.

Note that submodules are not initialized by default when cloning. Thus
you must either `git clone --recuse-submodules <URL>`, or after the
clone run `git submodules init`. Either of those commands will instruct
git to pull in the content from the submodules. Alternatively, there are
git-config options which may automate this task.