Split into columns, fix multi-line output
While beginning to implement a loop solution to our multiple line
output. I decided I wanted to break out one of the rows into two
columns. I felt this would make the loop easier to implement and format.
The intention was to split the columns, check output and work on the
loop.

In a very unexpected manner, the output was already split on newlines
for me. No need for loops! The change in code that allowed that was
moving from `append_tabbed` to `append`. Makes sense considering the
line in the library help page mentioning it's need for a loop. Just kind
of figured that comment applied to the other `append*` variations.

Either way, looks good, and it works.